Move vacuum constants to const.py
