+++ homeassistant/components/vacuum/const.py
@@ -2,8 +2,9 @@
 
 from __future__ import annotations
 
-from enum import StrEnum
+from enum import IntFlag, StrEnum
 from functools import partial
+from typing import TYPE_CHECKING
 
 from homeassistant.helpers.deprecation import (
     DeprecatedConstantEnum,
@@ -11,8 +12,15 @@ from homeassistant.helpers.deprecation import (
     check_if_deprecated_constant,
     dir_with_deprecated_constants,
 )
+from homeassistant.helpers.entity_component import EntityComponent
+from homeassistant.util.hass_dict import HassKey
+
+if TYPE_CHECKING:
+    from . import StateVacuumEntity
 
 DOMAIN = "vacuum"
+
+DATA_COMPONENT: HassKey[EntityComponent[StateVacuumEntity]] = HassKey(DOMAIN)
 
 
 class VacuumActivity(StrEnum):
@@ -24,6 +32,25 @@ class VacuumActivity(StrEnum):
     PAUSED = "paused"
     RETURNING = "returning"
     ERROR = "error"
+
+
+class VacuumEntityFeature(IntFlag):
+    """Supported features of the vacuum entity."""
+
+    TURN_ON = 1  # Deprecated, not supported by StateVacuumEntity
+    TURN_OFF = 2  # Deprecated, not supported by StateVacuumEntity
+    PAUSE = 4
+    STOP = 8
+    RETURN_HOME = 16
+    FAN_SPEED = 32
+    BATTERY = 64
+    STATUS = 128  # Deprecated, not supported by StateVacuumEntity
+    SEND_COMMAND = 256
+    LOCATE = 512
+    CLEAN_SPOT = 1024
+    MAP = 2048
+    STATE = 4096  # Must be set by vacuum platforms derived from StateVacuumEntity
+    START = 8192
 
 
 # These STATE_* constants are deprecated as of Home Assistant 2025.1.
